+++ lib/bootstrap-sass/engine.rb
@@ -2,7 +2,7 @@ module Bootstrap
   module Rails
     class Engine < ::Rails::Engine
       initializer "bootstrap-sass.assets.precompile" do |app|
-        app.config.assets.precompile += %w(glyphicons-halflings.png glyphicons-halflings-white.png)
+        app.config.assets.precompile << %r(bootstrap/glyphicons-halflings-regular\.(?:eot|svg|ttf|woff)$)
       end
     end
   end