+++ templates/project/manifest.rb
@@ -8,12 +8,6 @@ stylesheet '_variables.scss'
 # Other Bootstrap assets
 basedir = '../../vendor/assets'
 
-# Glyphicons sprites
-%w(glyphicons-halflings glyphicons-halflings-white).each do |file|
-  image "#{basedir}/images/#{file}.png", :to => "#{file}.png"
-end
-
-# Javascripts
-%w(affix alert button carousel collapse dropdown modal popover scrollspy tab tooltip transition typeahead).each do |file|
-  javascript "#{basedir}/javascripts/bootstrap-#{file}.js", :to => "bootstrap-#{file}.js"
+%w(affix alert button carousel collapse dropdown modal popover scrollspy tab tooltip transition).each do |file|
+  javascript "#{basedir}/javascripts/bootstrap/#{file}.js", :to => "bootstrap/#{file}.js"
 end